device: Add CaptureDevice struct
A capture device represents a Linux device node such as /dev/video0 with
type V4L2_BUF_TYPE_VIDEO_CAPTURE.

Other device types such as multi-planar or output ones shall get their
own struct in the device module.

At the moment, this struct provides a safe wrapper around the raw OS
file descriptor used for interacting with the device. The fd is closed in
the drop() impl of the struct.

There is explicit support for converting a DeviceInfo instance to a full
blown CaptureDevice one.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>