device: capture: Expose file descriptor
Remove the unused derived Debug/Default traits while we're at it.
The fd() function might go away in the future and the fd is only
exposed to other modules inside the crate.

Right now it's here to allow external implementations of VIDIOC calls on
a device created by library code.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>