v4l-rs: Add userptr buffer streaming support
mmap:

  (1) device --[MAP]--> kernel --[MAP]--> user
  (2) device --[DMA]--> kernel --[MAP]--> user

userptr:

  (3) device --[DMA]-->                   user

We now support two buffer transfer methods in this crate.

The first one, 'mmap' maps buffer into userspace. They may either be
backed by on-chip memory on the capture device itself (1), or allocated
in DMA-able main memory by the driver (2).

The second, 'userptr' directly has the driver do DMA transfers from the
device to main (application owned) memory. To create a continuous area of
memory in RAM suitable for DMA transfers, the driver may choose to swap
pages and lock them to prevent swapping out to disk. This is handled by
the virtual memory subsystem of the Linux kernel and is transparent to
the application.

References:

[mmap]    https://github.com/torvalds/linux/blob/master/Documentation/media/uapi/v4l/mmap.rst
[userptr] https://github.com/torvalds/linux/blob/master/Documentation/media/uapi/v4l/userp.rst

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>