buffers: Always execute VIDIOC_STREAM{ON|OFF}
From the v4l docs:

If VIDIOC_STREAMON is called when streaming is already in progress, or if
VIDIOC_STREAMOFF is called when streaming is already stopped, then 0 is
returned. Nothing happens in the case of VIDIOC_STREAMON, but VIDIOC_STREAMOFF
will return queued buffers to their starting state as mentioned above.

The last part is important to us: we want the driver to return the buffers to a
clean state when stopping the stream.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>