examples: Fix compressed format forwarding
The issue and its fix are described in the code.
Basically, we have to ensure that we always try to write exactly the
number of input bytes to the output buffer.

In Rust, the copy_from_slice(src) fn expects the src len to match that
of self. But for compressed formats, the output buffer would almost
always be larger than the input buffer.

Fixes https://github.com/raymanfx/libv4l-rs/issues/28.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>