io: arena: Replace unchecked access
Since we are not accessing elements in a hot loop, the unchecked
access does not benefit us much. Instead, it forces us to use
unsafe code.

Get rid of it by using index based array access. Worst case,
the code will panic which is the same as before and unexpected
behavior.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>